+++ front/src/components/Navbar/NavbarContext.tsx
@@ -1,0 +1,48 @@
+import React, { createContext, useContext, useState } from "react";
+import axios from "axios";
+
+const NavbarContext = createContext<any>(null);
+
+interface NavbarProviderProps {
+  children: React.ReactNode;
+}
+
+export const NavbarProvider: React.FC<NavbarProviderProps> = ({ children }) => {
+  const [user, setUser] = useState<any>(null);
+  const [loading, setLoading] = useState<boolean>(true);
+
+  const fetchUserData = async () => {
+    const accessToken = localStorage.getItem("accessToken");
+    if (!accessToken) {
+      setUser(null);
+      setLoading(false);
+      return;
+    }
+
+    try {
+      const response = await axios.get("/api/auth/me", {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      });
+      if (response.status === 200) {
+        setUser(response.data.user);
+      } else {
+        setUser(null);
+      }
+    } catch (error) {
+      console.error("Error fetching user data:", error);
+      setUser(null);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <NavbarContext.Provider value={{ user, fetchUserData, loading }}>
+      {children}
+    </NavbarContext.Provider>
+  );
+};
+
+export const useNavbar = () => useContext(NavbarContext);
